requirements.txt
pillow
torch
transformers
huggingface_hub
huggingface_hub[cli]